Modify the implementation of backpatch.
Students tend to be confused about backpatch. We change:
1. The prototype of function backPatch to void backPatch(std::vector<BasicBlock**> &list, BasicBlock*target)
2. The type of trueList and falseList to std::vector<BasicBlock**>
We set the element type of patch list to BasicBlock**, which is the address of `branch`, `true_branch`, `false_branch` in branch instruction.
So we can generate code for bool expression as follows:
new CmpInstruction(opcode, dst, src1, src2, bb);
if(gen_br)
{
    CondBrInstruction *cond;
    cond = new CondBrInstruction(nullptr, nullptr, dst, bb);    // the target basicblock is unknown.
    true_list.push_back(cond->truePatchBranch());
    false_list.push_back(cond->falsePatchBranch());
}